+++ src/redux/action-creators.js
@@ -575,6 +575,24 @@ export function getSinglePost(postId) {
   };
 }
 
+export function getPostForPreview(postId) {
+  return {
+    type: ActionTypes.GET_POST_FOR_PREVIEW,
+    apiRequest: Api.getPost,
+    nonAuthRequest: true,
+    payload: { postId },
+  };
+}
+
+export function getCommentForPreview(postId, commentId) {
+  return {
+    type: ActionTypes.GET_COMMENT_FOR_PREVIEW,
+    apiRequest: Api.getCommentForPreview,
+    nonAuthRequest: true,
+    payload: { postId, commentId },
+  };
+}
+
 export function getPostIdByOldName(oldName) {
   return {
     type: ActionTypes.GET_POST_ID_BY_OLD_NAME,